move memory into schedule (#4597)
