+++ examples/llm.c/export.py
@@ -5,9 +5,8 @@ from tinygrad import Device, nn, Tensor, dtypes, Variable
 Device.DEFAULT = "CLANG"
 from train_gpt2 import GPT, GPTConfig
 from tinygrad.helpers import dedup, to_function_name, flatten, getenv, GRAPH, GlobalCounters, ansilen, to_function_name
-from tinygrad.engine.schedule import create_schedule
+from tinygrad.engine.schedule import create_schedule, memory_planner
 from tinygrad.engine.realize import get_linearizer, run_schedule
-from tinygrad.engine.memory import memory_planner
 from tinygrad.ops import BufferOps, LoadOps
 
 TIMING = getenv("TIMING")
